+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **environment** arg to `JanusClient`.
 
 ## [3.57.1] - 2019-10-11
 ## Added
@@ -25,7 +27,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [3.55.2] - 2019-10-01
-
 ### Fixed
 - Warnings logged as errors